end tag omission on html, head and body
