+++ tests/html.test.ts
@@ -9,10 +9,6 @@ import {
   unreachable,
 } from "@std/assert";
 import {
-  attribute,
-  comment,
-  COMMENT_REGEX,
-  commentNode,
   customElementName,
   doctype,
   element,
@@ -23,250 +19,15 @@ import {
   isElementNode,
   isMNode,
   isTextNode,
-  NodeKind,
   serializeFragments,
-  serializeNode,
   shadowRoot,
-  spacesAndComments,
-  tagName,
   textNode,
 } from "../parser.ts";
-
-Deno.test("comments:regex", () => {
-  const r = new RegExp(COMMENT_REGEX.source + "$", "v");
-
-  // Can't start with >
-  assertEquals(r.test(">"), false);
-  // Can't start with ->
-  assertEquals(r.test("->"), false);
-  // Can't contain <!--
-  assertEquals(r.test("a<!--b"), false);
-  // Can't contain -->
-  assertEquals(r.test("a-->b"), false);
-  // Can't contain --!>
-  assertEquals(r.test("a--!>b"), false);
-  // Can't contain <!-
-  assertEquals(r.test("a<!-"), false);
-  // Can end in <!
-  assertEquals(r.test("a<!"), true);
-});
-
-Deno.test("comments:simple", () => {
-  const singleline = comment.parseOrThrow("<!-- A simple comment -->");
-  assertEquals(singleline, { kind: "COMMENT", text: " A simple comment " });
-
-  const multiline = comment.parseOrThrow(`<!--
-     A
-     multiline
-     comment
-     -->`);
-  assertEquals(multiline, {
-    kind: "COMMENT",
-    text: `
-     A
-     multiline
-     comment
-     `,
-  });
-
-  const edgeCaseComment = comment.parseOrThrow(`
-    <!--My favorite operators are > and <!-->
-  `.trim());
-
-  assertObjectMatch(edgeCaseComment, {
-    kind: NodeKind.COMMENT,
-    text: "My favorite operators are > and <!",
-  });
-});
-
-Deno.test("comments:complex", () => {
-  const consecutive_comments = spacesAndComments.parseOrThrow(`
-    <!-- consecutive comments -->
-    <!-- arrows ->-> -- > ->->->-- -> inside comments -->
-  `);
-  assertEquals(consecutive_comments, [
-    textNode("\n    "),
-    {
-      kind: "COMMENT",
-      text: " consecutive comments ",
-    },
-    {
-      kind: "COMMENT",
-      text: " arrows ->-> -- > ->->->-- -> inside comments ",
-    },
-    textNode("\n  "),
-  ]);
-
-  const html_inside_comment = element.parseOrThrow(`
-    <div><!-- <span>html inside comment</span> --></div>
-  `.trim());
-
-  assertObjectMatch(html_inside_comment, {
-    kind: ElementKind.NORMAL,
-    tagName: "div",
-    attributes: [],
-    children: [
-      { kind: "COMMENT", text: " <span>html inside comment</span> " },
-    ],
-  });
-});
-
-Deno.test("comments:nested", () => {
-  const nestedComments = fragments.parseOrThrow(`
-    <!-- This is a div -->
-
-    <div>
-
-      <!-- This is a p -->
-      <p>
-        Some text
-        <!-- This is a button -->
-        <button>click</button>
-        <!-- Now below the button -->
-      </p>
-
-      <!-- Another section -->
-
-      <!-- Another p -->
-      <p>
-        <input type="checkbox"> <!-- An input -->
-      </p>
-    </div>
-    `);
-
-  assertObjectMatch({ results: nestedComments }, {
-    results: [
-      textNode("\n    "),
-      commentNode(" This is a div "),
-      textNode("\n\n    "),
-      {
-        tagName: "div",
-        kind: ElementKind.NORMAL,
-        attributes: [],
-        children: [
-          textNode("\n\n      "),
-          commentNode(" This is a p "),
-          textNode("\n      "),
-          {
-            tagName: "p",
-            kind: ElementKind.NORMAL,
-            attributes: [],
-            children: [
-              { kind: "TEXT", text: "\n        Some text\n        " },
-              commentNode(" This is a button "),
-              textNode("\n        "),
-              {
-                tagName: "button",
-                kind: ElementKind.NORMAL,
-                attributes: [],
-                children: [{ kind: "TEXT", text: "click" }],
-              },
-              textNode("\n        "),
-              commentNode(" Now below the button "),
-              textNode("\n      "),
-            ],
-          },
-          textNode("\n\n      "),
-          commentNode(" Another section "),
-          textNode("\n\n      "),
-          commentNode(" Another p "),
-          textNode("\n      "),
-          {
-            tagName: "p",
-            kind: ElementKind.NORMAL,
-            attributes: [],
-            children: [
-              textNode("\n        "),
-              {
-                tagName: "input",
-                kind: ElementKind.VOID,
-                attributes: [["type", "checkbox"]],
-              },
-              textNode(" "),
-              commentNode(" An input "),
-              textNode("\n      "),
-            ],
-          },
-          textNode("\n    "),
-        ],
-      },
-      textNode("\n    "),
-    ],
-  });
-});
 
 Deno.test("doctype", () => {
   const res = doctype.parseOrThrow("<!Doctype Html >");
 
   assertEquals(res, textNode("<!DOCTYPE html>"));
-});
-
-Deno.test("attributes", () => {
-  const unquotedAttribute = attribute.parseOrThrow(`value=yes`);
-  assertEquals(unquotedAttribute, ["value", "yes"]);
-
-  const singleQuoteAttribute = attribute.parseOrThrow(`type='text'`);
-  assertEquals(singleQuoteAttribute, ["type", "text"]);
-
-  const doubleQuotesAttribute = attribute.parseOrThrow(`class="a b c"`);
-  assertEquals(doubleQuotesAttribute, ["class", "a b c"]);
-
-  const booleanAttribute = attribute.parseOrThrow(`checked`);
-  assertEquals(booleanAttribute, ["checked", ""]);
-
-  const nonAsciiAttribute = attribute.parseOrThrow(`xml:lang="us"`);
-  assertEquals(nonAsciiAttribute, ["xml:lang", "us"]);
-
-  const hangingBracket = element.parseOrThrow(
-    `<input
-    disabled
-    >`,
-  );
-  assertEquals(hangingBracket, {
-    tagName: "input",
-    kind: ElementKind.VOID,
-    attributes: [
-      [
-        "disabled",
-        "",
-      ],
-    ],
-  });
-
-  const recoverFromMissingWhiteSpace = element.parseOrThrow(
-    `<input value="yes"class="a b c">`,
-  );
-  assertEquals(recoverFromMissingWhiteSpace, {
-    tagName: "input",
-    kind: ElementKind.VOID,
-    attributes: [
-      ["value", "yes"],
-      ["class", "a b c"],
-    ],
-  });
-
-  const allowDuplicateAttributes = element.parseOrThrow(
-    `<input on:click="handleClick" on:click="log">`,
-  );
-  assertEquals(allowDuplicateAttributes, {
-    tagName: "input",
-    kind: ElementKind.VOID,
-    attributes: [
-      ["on:click", "handleClick"],
-      ["on:click", "log"],
-    ],
-  });
-
-  const keepAttributesCasing = element.parseOrThrow(
-    `<input prop:ariaChecked="checked">`,
-  );
-  assertEquals(keepAttributesCasing, {
-    tagName: "input",
-    kind: ElementKind.VOID,
-    attributes: [
-      ["prop:ariaChecked", "checked"],
-    ],
-  });
 });
 
 Deno.test("custom element names", () => {
@@ -287,84 +48,6 @@ Deno.test("custom element names", () => {
     assertInstanceOf(error, ParseError);
     assertStringIncludes(error?.message, "Forbidden custom element name");
   }
-});
-
-Deno.test("tag names", () => {
-  const name = tagName.parseOrThrow("Abc-d");
-  assertEquals(name, "abc-d");
-});
-
-Deno.test("disallow self-closing non-void element", () => {
-  try {
-    element.parseOrThrow("<div />");
-    unreachable();
-  } catch (error) {
-    assertInstanceOf(error, ParseError);
-    assert(
-      error.message.includes(
-        "Unexpected self-closing tag on a non-void element",
-      ),
-    );
-  }
-});
-
-Deno.test("void element", () => {
-  const input = element.parseOrThrow('<input type="text">');
-  assertEquals(input, {
-    tagName: "input",
-    kind: ElementKind.VOID,
-    attributes: [["type", "text"]],
-  });
-
-  // The closing slash becomes part of the unquoted attribute value
-  // https://html.spec.whatwg.org/#start-tags
-  const unquoted_attr_then_slash = element.parseOrThrow("<input type=text/>");
-  assertEquals(unquoted_attr_then_slash, {
-    tagName: "input",
-    kind: ElementKind.VOID,
-    attributes: [["type", "text/"]],
-  });
-});
-
-Deno.test("void elements shouldn't have a closing tag", () => {
-  try {
-    element.parseOrThrow('<input type="text"></input>');
-    unreachable();
-  } catch (error) {
-    assertInstanceOf(error, ParseError);
-    assert(error.message.includes("Unexpected end tag on a void element"));
-  }
-
-  try {
-    element.parseOrThrow(
-      `<input type="text">
-
-      </input>`,
-    );
-    unreachable();
-  } catch (error) {
-    assertInstanceOf(error, ParseError);
-    assert(error.message.includes("Unexpected end tag on a void element"));
-  }
-});
-
-Deno.test("void elements", () => {
-  const content = fragments.parseOrThrow(
-    '<img src="something.png"><br><input type=submit value=Ok />',
-  );
-  assertEquals(content, [
-    {
-      tagName: "img",
-      kind: ElementKind.VOID,
-      attributes: [["src", "something.png"]],
-    },
-    { tagName: "br", kind: ElementKind.VOID, attributes: [] },
-    {
-      tagName: "input",
-      kind: ElementKind.VOID,
-      attributes: [["type", "submit"], ["value", "Ok"]],
-    },
-  ]);
 });
 
 Deno.test("raw text elements", () => {
@@ -481,107 +164,6 @@ Deno.test("custom elements", () => {
 
   assertExists(res[0]);
   assertObjectMatch(res[0], node);
-});
-
-Deno.test("end tag omission ", () => {
-  const li_omissions = element.parseOrThrow(
-    `<ul>
-      <li>Apples
-      <li>Bananas
-    </ul>`,
-  );
-
-  assertEquals(
-    serializeNode(li_omissions),
-    `<ul>
-      <li>Apples
-      </li><li>Bananas
-    </li></ul>`,
-  );
-
-  const in_dl_omissions = element.parseOrThrow(
-    `<dl>
-      <dt>Coffee
-      <dd>Black hot drink
-      <dt>Milk
-      <dd>White cold drink
-    </dl>`,
-  );
-
-  assertEquals(
-    serializeNode(in_dl_omissions),
-    `<dl>
-      <dt>Coffee
-      </dt><dd>Black hot drink
-      </dd><dt>Milk
-      </dt><dd>White cold drink
-    </dd></dl>`,
-  );
-
-  const p_omissions = element.parseOrThrow(
-    `<body>
-      <p>This is the first paragraph.
-      <p>This is the second paragraph, and it ends when the next div begins.
-      <div>A block element</div>
-    </body>`,
-  );
-
-  assertEquals(
-    serializeNode(p_omissions),
-    `<body>
-      <p>This is the first paragraph.
-      </p><p>This is the second paragraph, and it ends when the next div begins.
-      </p><div>A block element</div>
-    </body>`,
-  );
-
-  const option_omissions = element.parseOrThrow(
-    `<select>
-      <option value="1">One
-      <option value="2">Two
-      <option value="3">Three
-    </select>`,
-  );
-
-  assertEquals(
-    serializeNode(option_omissions),
-    `<select>
-      <option value="1">One
-      </option><option value="2">Two
-      </option><option value="3">Three
-    </option></select>`,
-  );
-
-  const in_table_omissions = element.parseOrThrow(
-    `<table>
-  <caption>37547 TEE Electric Powered Rail Car Train Functions (Abbreviated)
-  <colgroup><col><col><col>
-  <thead>
-    <tr> <th>Function                              <th>Control Unit     <th>Central Station
-  <tbody>
-    <tr> <td>Headlights                            <td>✔                <td>✔
-    <tr> <td>Interior Lights                       <td>✔                <td>✔
-    <tr> <td>Electric locomotive operating sounds  <td>✔                <td>✔
-    <tr> <td>Engineer's cab lighting               <td>                 <td>✔
-    <tr> <td>Station Announcements - Swiss         <td>                 <td>✔
-  </table>`,
-  );
-
-  assertEquals(
-    serializeNode(in_table_omissions),
-    `<table>
-  <caption>37547 TEE Electric Powered Rail Car Train Functions (Abbreviated)
-  </caption><colgroup><col><col><col>
-  </colgroup><thead>
-    <tr> <th>Function                              </th><th>Control Unit     </th><th>Central Station
-  </th></tr></thead><tbody>
-    <tr> <td>Headlights                            </td><td>✔                </td><td>✔
-    </td></tr><tr> <td>Interior Lights                       </td><td>✔                </td><td>✔
-    </td></tr><tr> <td>Electric locomotive operating sounds  </td><td>✔                </td><td>✔
-    </td></tr><tr> <td>Engineer's cab lighting               </td><td>                 </td><td>✔
-    </td></tr><tr> <td>Station Announcements - Swiss         </td><td>                 </td><td>✔
-  </td></tr></tbody></table>`,
-  );
 });
 
 Deno.test("entities", () => {
